Tell Typescript about gen/ directory
Driver for this is to allow direct import of JS mojom bindings. This allows typescript to use strongly typed error checking for all generated mojom types, without having to add any typescript definition files. All the structs, enums, interfaces and union types are correctly inferred from the JS, as are the built-in modules.
The implementation in this PR is not the most ideal method. Improvements are mentioned in the TODOs, most notably generating a tsconfig.json inside the gen/ directory. However, this should work for most developer scenarios and we can perform the tsconfig.json in a quick follow-up.