Brave News: refactor to have API return card data and in intended display order, instead of a flat list of content
Previously, the front-end would decide on the order and type of card used for each feed item. Having the mojom API return that makes it easier to have the same decisions apply to desktop and android.
Also uses strong typings from mojom es module output. Works very well with Typescript.
Now that the mojom API is feature-complete, we can remove the extension API.