Store address in keyring meta
We need to be able to ask for permission for addresses even when the keyring is locked. To do this we store the address in the keyring meta info and we make `GetAccountInfosForKeyring` be independent of using an unlocked keyring.

We set the address info in meta when `ResumeDefaultKeyring` is called only for now to fix anyone that doesn’t have it set. The next time they unlock it will be set.